% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bjr.R
\name{bjr}
\alias{bjr}
\title{Bonhomme-Jochmans-Robin Estimator}
\usage{
bjr(
  y,
  II,
  K,
  basis = "Chebychev",
  cdf.points = NULL,
  estimate.mean.sd = TRUE,
  estimate.cdf = TRUE,
  estimate.mixing.prop = TRUE,
  use.Rust = FALSE,
  debug.return = "never",
  debug.output.to.csv = FALSE,
  control = list(cluster.threads = NULL)
)
}
\arguments{
\item{y}{Data}

\item{II}{Number of evaluation points for basis function}

\item{K}{Number of distribution components}

\item{basis}{Choice of basis function}

\item{cdf.points}{Points to evaluate the estimated CDF}

\item{estimate.mean.sd}{Estimate mean and sd?}

\item{estimate.cdf}{Estimate CDF?}

\item{estimate.mixing.prop}{Estimate mixing proportions?}

\item{use.Rust}{Use Rust implementation of certain sub-functions?}

\item{debug.return}{Return function workspace objects before end of function.
Supported values include "before est.AA.inner", "after est.AA.inner",
"before est.cdf.and.mixing.prop.inner", and "after est.cdf.and.mixing.prop.inner"}

\item{debug.output.to.csv}{TRUE/FALSE to output objects to csv when
debug.return is enabled. A directory above the current working directory
called "decoyanalysis-debug-output" will be created to contain the csv files.}

\item{control}{control list}
}
\value{
The results
}
\description{
Bonhomme-Jochmans-Robin Estimator
}
\examples{
\dontrun{
N = 500 # Sample size
M = 3   # Number of variables (repeated measurements) 
K = 2   # Number of distribution components
theta1 = c(0.0, 2.0) # Mean parameters
theta2 = c(1.0, 1.0) # Standard deviation parameters
omega  = c(0.3, 0.7) # Mixing proportions

set.seed(314)
y <- gen.bjr.normal.data(N = N, M = M, K = K, theta1 = theta1,
  theta2 = theta2, omega = omega)
bjr.results <- bjr(y, II = 10, K = 2)
}
}
