% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monte-carlo.R
\name{gen.bjr.normal.data}
\alias{gen.bjr.normal.data}
\title{Title}
\usage{
gen.bjr.normal.data(N, T, K, theta1, theta2, omega)
}
\arguments{
\item{N}{Sample size}

\item{T}{Number of variables (repeated measurements)}

\item{K}{Number of components}

\item{theta1}{Vector of mean parameters}

\item{theta2}{Vector of standard deviation parameters}

\item{omega}{Vector of mixing proportions}
}
\value{
N x T matrix
}
\description{
Title
}
\examples{
N = 500 # Sample size
T = 3   # Number of variables (repeated measurements) 
K = 2   # Number of components
theta1 = c(0.00, 3.00) # Mean parameters
theta2 = c(1.00, 1.00) # Standard deviation parameters
omega  = c(0.30, 0.70) # Mixing proportions

set.seed(314)
y <- gen.bjr.normal.data(N = N, T = T, K = K, theta1 = theta1, theta2 = theta2, omega = omega)
}
