% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monte-carlo.R
\name{gen.old.new.dsa.distributions}
\alias{gen.old.new.dsa.distributions}
\title{Create functions to draw random values from old and new DSAs and real spend distribution}
\usage{
gen.old.new.dsa.distributions()
}
\value{
A list with 3 elements: \code{draw.decoy.old.dsa}, \code{draw.decoy.new.dsa},
and \code{draw.real.spend}. These elements are functions that take a single
argument: the number of random values to be drawn.
}
\description{
Create functions to draw random values from old and new DSAs and real spend distribution
}
\details{
Creates functions to draw random values from the old and new Decoy Selection
Algorithms (DSAs) and the real spend distribution.
}
\examples{
1

}
