% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monte-carlo.R
\name{gen.old.new.dsa.rings}
\alias{gen.old.new.dsa.rings}
\title{Generate rings from old and new DSAs and their antecedents}
\usage{
gen.old.new.dsa.rings(
  n.rings,
  beta,
  old.dsa,
  new.dsa,
  real.spend,
  ring.size = 16,
  change.probability = 0.4,
  check.data.sorted = FALSE
)
}
\arguments{
\item{n.rings}{Number of rings to generate}

\item{beta}{Share of rings that are created by the new DSA}

\item{old.dsa}{Function that draws a random variable from the old DSA. It
should have a single argument: the number of values to draw.}

\item{new.dsa}{Function that draws a random variable from the new DSA.}

\item{real.spend}{Function that draws a random variable from the real spend
distribution.}

\item{ring.size}{Ring size. 16 by default.}

\item{change.probability}{Probability that a ring spends the change in the
wallet from a prior transaction.}

\item{check.data.sorted}{Double-check if the ring data is properly sorted?}
}
\value{
A list with three elements:

\code{X}: a matrix with \code{n.rings} rows and \code{ring.size + ring.size^2} columns of
the output ages of each ring member and its antecedents. The column
"Inputs.0.Previous_Tx_Block_Num_Delta.self.i" is the age of the \code{i}th ring
member of the transaction of interest.
"Inputs.0.Previous_Tx_Block_Num_Delta.j.k" is the age of the \code{k}th
ring member of the antecedent transaction referenced in the \code{j}th ring
member slot of the transaction of interest.

\code{y}: a vector of length \code{n.rings} containing the ring member position of
the real spend, indexed from zero.

\code{z}: a vector of length \code{n.rings} with zero where the ring is an old-DSA
ring and one where the ring is a new-DSA ring.
}
\description{
Generate rings from old and new DSAs and their antecedents
}
\details{
For more information on variable definitions and the behavioral model,
see "Formula for Accuracy of Guessing Monero Real Spends Using Fungibility
Defects":
\url{https://github.com/Rucknium/misc-research/blob/main/Monero-Fungibility-Defect-Classifier/pdf/classify-real-spend-with-fungibility-defects.pdf}
}
\examples{
\dontrun{
distributions <- gen.old.new.dsa.distributions()

ring.data <- gen.old.new.dsa.rings(n.rings = 10000, beta = 0.10,
  old.dsa = distributions$draw.decoy.old.dsa,
  new.dsa = distributions$draw.decoy.new.dsa,
  real.spend = distributions$draw.real.spend)

write.csv(ring.data$X, file = "X.csv", row.names = FALSE)
cat(ring.data$y, file = "y.csv", sep = "\n")
cat(ring.data$z, file = "z.csv", sep = "\n")
}
}
