% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monte-carlo.R
\name{gen.ring.data}
\alias{gen.ring.data}
\title{Title}
\usage{
gen.ring.data(
  n.rings,
  ring.size,
  mixing.proportions,
  decoy.dists,
  real.spend.dists
)
}
\arguments{
\item{n.rings}{S}

\item{ring.size}{S}

\item{mixing.proportions}{S}

\item{decoy.dists}{S}

\item{real.spend.dists}{S}
}
\value{
Matrix
}
\description{
Title
}
\examples{
1

}
