% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-decoder.R
\name{map.decoder.success.prob}
\alias{map.decoder.success.prob}
\title{MAP Decoder Effectiveness}
\usage{
map.decoder.success.prob(
  f_S,
  f_D,
  n.decoys,
  subsample = 1,
  rounding.digit = NULL,
  return.only.success.probability.vector = FALSE
)
}
\arguments{
\item{f_S}{Probability mass function (PMF) of the real spend distribution.
This should be a vector of probabilities summing to one.}

\item{f_D}{PMF of the decoy distribution. Must be the same length as \code{f_S}.}

\item{n.decoys}{Number of decoys.}

\item{subsample}{A number between 0 and 1. Takes a subsample of the support
of \code{f_S}, as a proportion, to speed up computation. The subsample is generated
by weighted random sampling without replacement, using the implementation of
\code{wrswoR::sample_int_expj()}. Recommended not to go below 0.1. Alternatively,
a vector of support points. This alternative specification can be used
if the user desires to specify the support subsample directly.}

\item{rounding.digit}{Optional. If specified, round an intermediate
vector to the nearest digit. Helps speed up computation, with minimal
loss of accuracy of the final result. Setting to 5 seems to work well.}

\item{return.only.success.probability.vector}{If TRUE, returns only the
\code{success.probability.vector} object}
}
\value{
A list with the following elements:
\itemize{
\item \code{mean.success.probability}: A scalar, between 0 and 1. The value of
\eqn{L_{MAP\:Decoder}}.
\item \code{success.probability.vector}: A vector of the same length as \code{sub.supp}.
This vector is the computation of the expression inside the parentheses of
the formula that is taken to the power \eqn{M}. Note that this vector is
"raw" and has not yet been taken to the power \eqn{M}. This vector is useful
for analyzing which output ages are most vulnerable to the MAP Decoder
attack.
\item \code{sub.supp}: The subsample of the support. If the \code{subsample} argument is
left at its default value of 1, then this is just a vector of the integers
from 1 to the length of \code{f_S}.
If \code{return.only.success.probability.vector} is specified as TRUE, only
the \code{success.probability.vector} vector is returned, without a list.
}
}
\description{
MAP Decoder Effectiveness
}
\details{
Computes equation 13 of Aeeneh et al. (2021).

Let \eqn{Z} be the total number of outputs eligible to be spent in a RingCT
ring. Let \eqn{f_S} be the real spend probability mass function (PMF). Let
\eqn{f_D} be a proposed decoy PMF. Let \eqn{M} be the number of decoys in a
ring. Let \eqn{\mathbf{1}\{\}} be the indicator function that takes value 1
when the statement within the braces is true and 0 otherwise. The average
success probability of the MAP Decoder attack is
\deqn{L_{MAP\:Decoder}=\sum_{i=1}^{Z}f_{S}\left(i\right)\left(\sum_{j=1}^{Z}f_{D}\left(j\right)\mathbf{1}\left\{ \tfrac{f_{S}\left(j\right)}{f_{D}\left(j\right)}<\tfrac{f_{S}\left(i\right)}{f_{D}\left(i\right)}\right\} \right)^{M}}
}
\examples{
n.outputs <- 1000

real.spend <- dnbinom(seq_len(n.outputs), size = 1, mu = 500)
decoy <- dnbinom(seq_len(n.outputs), size = 5, mu = 500)
# Negative binomial distributions

real.spend <- real.spend/sum(real.spend)
decoy <- decoy/sum(decoy)

plot(decoy, type = "l", ylim = c(0, max(c(real.spend, decoy))),
  ylab = "Probability mass function")
lines(real.spend, col = "red")
legend("topright", legend = c("Decoy", "Real spend"),
  col = c("black", "red"), lty = 1)
  
result <- map.decoder.success.prob(f_S = real.spend, f_D = decoy,
  n.decoys = 15)

print(result$mean.success.probability)

}
