% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patra-sen.R
\name{patra.sen.bjr}
\alias{patra.sen.bjr}
\title{Title}
\usage{
patra.sen.bjr(Fn.hat.value, supp.points, Fb, alpha)
}
\arguments{
\item{Fn.hat.value}{Estimated CDF from bjr()}

\item{supp.points}{Points on the support corresponding to Fn.hat.value}

\item{Fb}{A cumulative density function such as pnorm()}

\item{alpha}{The "contaminant's" proportion of the mixture distribution}
}
\value{
An object
}
\description{
Title
}
\examples{
1
}
